refactor: crate::rpc_api::data_types -> crate::rpc::types
