+++ src/rpc/state_api.rs
@@ -6,8 +6,8 @@ use crate::cid_collections::CidHashSet;
 use crate::libp2p::NetworkMessage;
 use crate::lotus_json::LotusJson;
 use crate::rpc::error::JsonRpcError;
+use crate::rpc::types::*;
 use crate::rpc::Ctx;
-use crate::rpc_api::data_types::*;
 use crate::shim::{
     address::Address, clock::ChainEpoch, deal::DealID, econ::TokenAmount, executor::Receipt,
     state_tree::ActorState, version::NetworkVersion,
