+++ src/rpc/sync_api.rs
@@ -5,8 +5,8 @@
 use crate::chain_sync::SyncState;
 use crate::lotus_json::LotusJson;
 use crate::rpc::error::JsonRpcError;
+use crate::rpc::types::RPCSyncState;
 use crate::rpc::Ctx;
-use crate::rpc_api::data_types::RPCSyncState;
 
 use anyhow::Result;
 use fvm_ipld_blockstore::Blockstore;
