+++ src/rpc_client/common_ops.rs
@@ -1,10 +1,8 @@
 // Copyright 2019-2024 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::rpc_api::{
-    common_api::{DISCOVER, SESSION, SHUTDOWN, START_TIME, VERSION},
-    data_types::{APIVersion, DiscoverResult},
-};
+use crate::rpc::types::{APIVersion, DiscoverResult};
+use crate::rpc_api::common_api::{DISCOVER, SESSION, SHUTDOWN, START_TIME, VERSION};
 use chrono::{DateTime, Utc};
 
 use super::{ApiInfo, JsonRpcError, RpcRequest};
