+++ src/state_manager/utils.rs
@@ -244,9 +244,7 @@ mod test {
 
 /// Parsed tree of [`fvm4::trace::ExecutionEvent`]s
 pub mod structured {
-    use crate::rpc_api::data_types::{
-        ActorTrace, ExecutionTrace, GasTrace, MessageTrace, ReturnTrace,
-    };
+    use crate::rpc::types::{ActorTrace, ExecutionTrace, GasTrace, MessageTrace, ReturnTrace};
     use std::collections::VecDeque;
 
     use crate::shim::{
